+++ include/fluxgl/graphics/renderer.h
@@ -2,6 +2,7 @@
 
 #include <glm/glm.hpp>
 #include <fluxgl/components/renderable.h>
+#include <fluxgl/components/camera.h>
 
 namespace fluxgl {
     class Renderer {
@@ -11,6 +12,6 @@ namespace fluxgl {
 
         public:
             static void clear(const glm::vec3& color = {0.1f, 0.1f, 0.1f});
-            static void draw(const Renderable& renderable);
+            static void draw(const Renderable& renderable, const Camera* camera = nullptr);
     };
 }