+++ src/graphics/renderer.cpp
@@ -8,7 +8,7 @@ namespace fluxgl {
         glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
     }
 
-    void Renderer::draw(const Renderable& renderable) {
+    void Renderer::draw(const Renderable& renderable, const Camera* camera) {
         const Material& material = renderable.material;
         const Mesh& mesh = renderable.mesh;
         const Transform& transform = renderable.transform;
@@ -61,9 +61,14 @@ namespace fluxgl {
         material.shader.setUniform("u_Roughness", material.roughness);
         material.shader.setUniform("u_AmbientOcclusion", material.ambientOcclusion);
 
-        // Binding model matrix
+        // Binding MVP matrices
         glm::mat4 model = transform.getModelMatrix();
-        material.shader.setUniform("model", model);
+        glm::mat4 view = camera ? camera->getViewMatrix() : glm::mat4(1.0f);
+        glm::mat4 projection = camera ? camera->getProjectionMatrix() : glm::mat4(1.0f);
+
+        material.shader.setUniform("u_View", view);
+        material.shader.setUniform("u_Projection", projection);
+        material.shader.setUniform("u_Model", model);
 
         // DRAW CALL
         unsigned int vao = mesh.getVAO();